lsh first implementation

+++ src/lsh/Element.java
@@ -2,27 +2,27 @@ package lsh;
 
 public class Element {
 
-	private Double value;
+	private int value;
 	
 	private boolean infinity;
 	
 	public Element() {
 		this.infinity = true;
 	}
-	
-	public Element(Double value){
+
+	public Element(int value){
 		setValue(value);
 		setInfinity(false);
 	}
-
-	public Double getValue() {
+	
+	public int getValue() {
 		if(isInfinity())
 			throw new NullPointerException();
 		
 		return value;
 	}
 
-	public void setValue(Double value) {
+	public void setValue(int value) {
 		if(isInfinity())
 			setInfinity(false);
